Requested changes to branch for pull request

+++ src/Controller/Physics/PhysicBody.cpp
@@ -2,11 +2,11 @@
 
 using namespace rp3d;
 
-int PhysicsBody::colliderSize() { 
+size_t PhysicsBody::colliderSize() {
 	return colliders.size();
 }
 
-void PhysicsBody::removeCollider(int index)
+void PhysicsBody::removeCollider(size_t index)
 {
 	if (index < 0 || index >= colliders.size())
 		return;
@@ -51,7 +51,7 @@ bool PhysicsBody::getIsTrigger()
 	return is_trigger;
 }
 
-glm::vec3 PhysicsBody::getColliderPosition(int index, Apply type)
+glm::vec3 PhysicsBody::getColliderPosition(size_t index, Apply type)
 {
 	Vector3 temp;
 	switch (type)
@@ -68,7 +68,7 @@ glm::vec3 PhysicsBody::getColliderPosition(int index, Apply type)
 	return glm::vec3(temp.x, temp.y, temp.z);
 }
 
-glm::vec4 PhysicsBody::getColliderOrientation(int index, Apply type)
+glm::vec4 PhysicsBody::getColliderOrientation(size_t index, Apply type)
 {
 	Quaternion temp;
 
@@ -86,17 +86,17 @@ glm::vec4 PhysicsBody::getColliderOrientation(int index, Apply type)
 	return glm::vec4(temp.w, temp.x, temp.y, temp.z);
 }
 
-float PhysicsBody::getColliderBounce(int index){
+float PhysicsBody::getColliderBounce(size_t index){
 	return colliders[index]->getMaterial().getBounciness();
 }
-float PhysicsBody::getColliderFriction(int index){
+float PhysicsBody::getColliderFriction(size_t index){
 	return colliders[index]->getMaterial().getFrictionCoefficient();
 }
-float PhysicsBody::getColliderMassDesity(int index) {
+float PhysicsBody::getColliderMassDesity(size_t index) {
 	return colliders[index]->getMaterial().getMassDensity();
 }
 
-int PhysicsBody::getColliderType(int index)
+int PhysicsBody::getColliderType(size_t index)
 {
 	switch (colliders[index]->getCollisionShape()->getName())
 	{
@@ -111,7 +111,7 @@ int PhysicsBody::getColliderType(int index)
 	}
 }
 
-void PhysicsBody::addMaterialToCollider(int index, float bounce,
+void PhysicsBody::addMaterialToCollider(size_t index, float bounce,
 	float mass_density, float friction)
 {
 	Material& material = colliders[index]->getMaterial();
@@ -121,7 +121,8 @@ void PhysicsBody::addMaterialToCollider(int index, float bounce,
 }
 
 
-const BoxShape* PhysicsBody::getColliderBox(int index){
+const BoxShape* PhysicsBody::getColliderBox(size_t index){
+
 	try {
 		if (getColliderType(index) == 3)
 			return m_box.find(colliders[index])->second;
@@ -133,7 +134,7 @@ const BoxShape* PhysicsBody::getColliderBox(int index){
 		std::cout << "ERROR: " << err << std::endl;
 	}
 }
-const SphereShape* PhysicsBody::getColliderSphere(int index) {
+const SphereShape* PhysicsBody::getColliderSphere(size_t index) {
 	try {
 		if (getColliderType(index) == 1)
 			return m_sphere.find(colliders[index])->second;
@@ -145,7 +146,7 @@ const SphereShape* PhysicsBody::getColliderSphere(int index) {
 		std::cout << "ERROR: " << err << std::endl;
 	}
 }
-const CapsuleShape* PhysicsBody::getColliderCapsule(int index) {
+const CapsuleShape* PhysicsBody::getColliderCapsule(size_t index) {
 	try {
 		if (getColliderType(index) == 2)
 			return m_capsule.find(colliders[index])->second;
